Connect Tools to editor slice
