+++ ui/src/components/editor/PlateSidebar/Tools/Tools.js
@@ -1,22 +1,17 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 import { Menu, Icon } from 'semantic-ui-react';
 
-import {
-  setClickMode,
-  deselectAllWells,
-} from '../../../../store/designActions';
-import { selectActivePlate } from '../../../../store/selectors';
 import { Header } from './Header';
 import { ApplyTool } from './ApplyTool';
 import { ClearTool } from './ClearTool';
 import styles from './Tools.module.css';
 
-class Tools extends Component {
+export class Tools extends Component {
   handleToolClick = (e, { name }) => {
-    this.props.setClickMode(name);
-    this.props.deselectAllWells({ plateId: this.props.activePlate.id });
+    if (this.props.onToolMenuItemClick) {
+      this.props.onToolMenuItemClick({ clickMode: name });
+    }
   };
   renderTool() {
     const { clickMode } = this.props;
@@ -66,25 +61,6 @@ class Tools extends Component {
 }
 
 Tools.propTypes = {
-  activePlate: PropTypes.object,
   clickMode: PropTypes.string.isRequired,
-  setClickMode: PropTypes.func.isRequired,
-  deselectAllWells: PropTypes.func.isRequired,
+  onToolMenuItemClick: PropTypes.func,
 };
-
-const mapState = (state, props) => {
-  const activePlate = selectActivePlate(state);
-  const { clickMode } = state.designExperiment;
-  return { activePlate, clickMode };
-};
-
-const mapDispatch = {
-  setClickMode,
-  deselectAllWells,
-};
-
-const connected = connect(
-  mapState,
-  mapDispatch
-)(Tools);
-export { connected as Tools };
